minimizes vscode-extension package to export deps only
