+++ .changeset/few-jobs-hammer.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-vscode-extension-dependencies': major
+---
+
+minimizes vscode-extension package to export deps only
