--- packages/legend-vscode-extension-dependencies/src/utils/Const.ts
@@ -1,21 +1,0 @@
-/**
- * Copyright (c) 2020-present, Goldman Sachs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// Event Types
-export const GET_PROJECT_ENTITIES = 'getProjectEntities';
-export const GET_PROJECT_ENTITIES_RESPONSE = 'getProjectEntitiesResponse';
-export const DIAGRAM_DROP_CLASS_ERROR = 'diagramDropClassError';
-export const WRITE_ENTITY = 'writeEntity';
