+++ packages/legend-vscode-extension-dependencies/style/index.scss
@@ -17,16 +17,3 @@
 @import url('@finos/legend-extension-dsl-diagram/lib/index.css');
 @import url('@finos/legend-art/lib/normalize.css');
 @import url('@finos/legend-art/lib/index.css');
-
-.diagram-editor__error {
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  &__details {
-    border-bottom: 1px dotted white;
-    cursor: help;
-    margin-left: 5px;
-  }
-}
